requirements.txt
# --- AI (Claude Computer Use) ---
anthropic>=0.43.0

# --- Browser Automation ---
playwright>=1.44.0
Pillow>=10.2.0

# --- Server ---
fastapi
uvicorn
pydantic
python-dotenv
python-multipart
tenacity

# --- TTS & STT ---
httpx>=0.27.0
